+++ Basil/include/Scene/Component.h
@@ -3,10 +3,22 @@
  */
 #pragma once
 
+#include "Renderer/Camera.h"
+
 #include <glm/glm.hpp>
+#include <glm/gtc/matrix_transform.hpp>
 
 namespace Basil
 {
+	struct CameraComponent
+	{
+		CameraComponent() = default;
+		CameraComponent(const CameraComponent&) = default;
+		CameraComponent(const glm::mat4& projection) : camera(projection) {}
+		Camera camera;
+		bool primary = true;
+	};
+
 	struct SpriteRenderComponent
 	{
 		SpriteRenderComponent() = default;
@@ -28,6 +40,7 @@ namespace Basil
 		TransformComponent() = default;
 		TransformComponent(const TransformComponent&) = default;
 		TransformComponent(const glm::mat4& transform) : transform(transform) {}
+		const void updateTransform() { transform = glm::translate(glm::mat4(1.0f), position) * glm::rotate(glm::mat4(1.0f), glm::radians(rotation), glm::vec3(0.0f, 0.0f, 1.0f)) * glm::scale(glm::mat4(1.0f), scale); }
 		operator glm::mat4& () { return transform; }
 		operator const glm::mat4& () const { return transform; }
 		glm::mat4 transform{ 1.0f };